Moved relevant files into new folder
